+++ src/modules/index.ts
@@ -6,6 +6,8 @@ export * from "./offset";
 export * from "./tokenization";
 export * from "./getCharacteristicsBytes";
 export * from "./fungibleTrading";
+export * from "./rollTokens";
 export * from "./NFTtrading/getTokenList";
 export * from "./NFTtrading/orderbook";
 export * from "./NFTtrading/NFTTrading";
+export * from "./carbonInfo";