+++ test/init/thea.spec.ts
@@ -1,4 +1,5 @@
 import {
+	CarbonInfo,
 	Convert,
 	FungibleTrading,
 	GetCharacteristicsBytes,
@@ -12,14 +13,22 @@ import {
 	Unwrap
 } from "../../src";
 import { consts, TheaError } from "../../src/utils";
-import { ExternalProvider, InfuraProvider, Provider, Web3Provider } from "@ethersproject/providers";
+import { ExternalProvider, InfuraProvider, Network, Provider, Web3Provider } from "@ethersproject/providers";
 import { Wallet } from "@ethersproject/wallet";
 import { PRIVATE_KEY } from "../mocks";
 import * as utils from "../../src/utils/utils";
+import { Signer } from "@ethersproject/abstract-signer";
 jest.mock("../../src/modules/");
+/* eslint-disable  @typescript-eslint/no-explicit-any */
 jest.mock("../../src/utils/utils", () => {
 	return {
-		getCurrentNBTTokenAddress: jest.fn().mockResolvedValue("0x5FbDB2315678afecb367f032d93F642f64180aa3")
+		getCurrentNBTTokenAddress: jest.fn().mockResolvedValue("0x5FbDB2315678afecb367f032d93F642f64180aa3"),
+		isSigner(providerOrSigner: any): providerOrSigner is Signer {
+			return !!providerOrSigner._isSigner;
+		},
+		isProvider(providerOrSigner: any): providerOrSigner is Provider {
+			return !!providerOrSigner._isProvider;
+		}
 	};
 });
 jest.mock("@ethersproject/providers", () => {
@@ -27,12 +36,22 @@ jest.mock("@ethersproject/providers", () => {
 		Web3Provider: jest.fn().mockImplementation(() => {
 			return {
 				getSigner: jest.fn().mockImplementation(() => {
-					return {};
+					return {
+						_isSigner: true,
+						getChainId: (): Promise<number> => {
+							return Promise.resolve(80001);
+						}
+					};
 				})
 			};
 		}),
-		InfuraProvider: jest.fn().mockImplementation(() => {
-			const value = { _isProvider: true };
+		InfuraProvider: jest.fn().mockImplementation((args) => {
+			const value = {
+				_isProvider: true,
+				getNetwork: (): Promise<Network> => {
+					return Promise.resolve({ chainId: args ?? 80001, name: "Mumbai" });
+				}
+			};
 			return value as Provider;
 		})
 	};
@@ -85,12 +104,22 @@ describe("TheaSDK", () => {
 			expect(Orderbook).toBeCalled();
 			expect(NFTTrading).toBeCalled();
 			expect(GetTokenList).toBeCalled();
+			expect(CarbonInfo).toBeCalled();
 			expect(currentNbtSpy).toBeCalled();
 			expect(consts[TheaNetwork.MUMBAI].currentNbtTokenContract).toBe("0x5FbDB2315678afecb367f032d93F642f64180aa3");
 		});
 
-		it("should return TheaSDK instance using signer", async () => {
+		it("should throw error if signer doesn't have provider specified", async () => {
 			const signer = new Wallet(PRIVATE_KEY);
+			await expect(
+				TheaSDK.init({
+					network: TheaNetwork.MUMBAI,
+					signer
+				})
+			).rejects.toThrow(new TheaError({ type: "SIGNER_REQUIRES_PROVIDER", message: "Signer must be have provider" }));
+		});
+		it("should return TheaSDK instance using signer", async () => {
+			const signer = new Wallet(PRIVATE_KEY, new InfuraProvider());
 			const result = await TheaSDK.init({
 				network: TheaNetwork.MUMBAI,
 				signer
@@ -122,6 +151,22 @@ describe("TheaSDK", () => {
 			expect(result).toBeInstanceOf(TheaSDK);
 			expect(result.network).toBe(TheaNetwork.MUMBAI);
 			expect(result.providerOrSigner).toBeDefined();
+		});
+
+		it("should throw error if there is network mismatch between provider and network in init options", async () => {
+			const provider = new InfuraProvider(1);
+			const { chainId } = await provider.getNetwork();
+			const options = {
+				network: TheaNetwork.MUMBAI,
+				provider
+			};
+
+			await expect(TheaSDK.init(options)).rejects.toThrow(
+				new TheaError({
+					type: "NETWORK_MISMATCH",
+					message: `Provided network is ${options.network} but provider is connected to ${chainId} network`
+				})
+			);
 		});
 	});
 });